add hotels, rooms and reservation tables

+++ src/main/resources/db/migration/V2__rooms_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE rooms (
-  id SERIAL PRIMARY KEY,
-  hotel_id INTEGER NOT NULL REFERENCES hotels (id),
+  id BIGSERIAL PRIMARY KEY,
+  hotel_id BIGINT NOT NULL REFERENCES hotels (id),
   room_number VARCHAR(10),
   room_type VARCHAR(50),
   price DECIMAL(10, 2) NOT NULL,
